Improve packet parsing.
Mostly, handle byte stuffing properly.

Once we determine that a byte can't be the start of a data item, we
skip over it as part of the packet parsing instead of waiting looping
over all bytes looking for the start-of-data indicator.

Now that we shouldn't see a start-of-data indicator in the data, if we
do, we ignore the data item being parsed and start over at this point.